% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_metric.R
\name{derive_metric}
\alias{derive_metric}
\title{Derive tidycyte metric using arithmetic or function}
\usage{
derive_metric(.df, .name, .formula)
}
\arguments{
\item{.df}{A data frame containing tidycyte data.}

\item{.name}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> The masked name of the metric to be derived (can overwrite existing metrics)}

\item{.formula}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> A formula containing masked metric names. May use arithmetic or other R functions, and can make use of other metrics.}
}
\value{
An updated data frame containing the new or revised metric.
}
\description{
Permits the use of arithmetic or other R functions for each metric. Can accept other metrics in arguments.
}
\examples{
df \%>\% derive_metric(Yellow, sqrt(`h2-3`) / mAzalea^2 - 3) # mathematical operations and offsets
df \%>\% derive_metric(mAzalea, mAzalea - mean(mAzalea)) # can use other functions
df \%>\% derive_metric(mAzalea, mAzalea - as.numeric(quantile(mAzalea,0.05))) # minus bottom 5\%ile
df \%>\% derive_metric(mAzalea_scaled, mAzalea) \%>\% # copy to new name rather than the original name
  scale_metric(mAzalea_scaled) # rescale from 0 to 1 by default, returns the same name
}
\keyword{function}
\keyword{metric}
\keyword{tidycyte}
