% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_to_cell.R
\name{id_to_cell}
\alias{id_to_cell}
\title{Extracts cell name from tidycyte sample ID}
\usage{
id_to_cell(id)
}
\arguments{
\item{id}{A character string or vector of character strings formatted as an Incucyte sample ID.}
}
\value{
A character string or vector of character strings containing the cell name from each sample ID.
}
\description{
Utility function that returns the cell name for each id.
}
\examples{
df$cell <- id_to_cell(df$id)
}
\keyword{cell}
\keyword{id}
\keyword{tidycyte}
