% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_to_image.R
\name{id_to_image}
\alias{id_to_image}
\title{Extracts image number from tidycyte sample ID}
\usage{
id_to_image(id)
}
\arguments{
\item{id}{A character string or vector of character strings formatted as an Incucyte sample ID.}
}
\value{
A character string or vector of character strings containing the image number from each sample ID.
}
\description{
Utility function that returns the image number for each id.
}
\examples{
df$image <- id_to_image(df$id)
}
\keyword{id}
\keyword{image}
\keyword{tidycyte}
