% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_to_passage.R
\name{id_to_passage}
\alias{id_to_passage}
\title{Extracts passage number from tidycyte sample ID}
\usage{
id_to_passage(id)
}
\arguments{
\item{id}{A character string or vector of character strings formatted as an Incucyte sample ID.}
}
\value{
A character string or vector of character strings containing the passage number from each sample ID.
}
\description{
Utility function that returns the passage number for each id.
}
\examples{
df$passage <- id_to_passage(df$id)
}
\keyword{id}
\keyword{passage}
\keyword{tidycyte}
