% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_to_seeding.R
\name{id_to_seeding}
\alias{id_to_seeding}
\title{Extracts seeding conditions from tidycyte sample ID}
\usage{
id_to_seeding(id)
}
\arguments{
\item{id}{A character string or vector of character strings formatted as an Incucyte sample ID.}
}
\value{
A character string or vector of character strings containing the seeding conditions from each sample ID.
}
\description{
Utility function that returns seeding conditions for each id.
}
\examples{
df$seeding <- id_to_seeding(df$id)
}
\keyword{id}
\keyword{passage}
\keyword{tidycyte}
