% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_to_treatment.R
\name{id_to_treatment}
\alias{id_to_treatment}
\title{Extracts treatment from tidycyte sample ID}
\usage{
id_to_treatment(id)
}
\arguments{
\item{id}{A character string or vector of character strings formatted as an Incucyte sample ID.}
}
\value{
A character string or vector of character strings containing the treatment from each sample ID.
}
\description{
Utility function that returns the treatment for each id.
}
\examples{
df$treatment <- id_to_treatment(df$id)
}
\keyword{id}
\keyword{tidycyte}
\keyword{treatment}
