% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_to_well.R
\name{id_to_well}
\alias{id_to_well}
\title{Extracts well from tidycyte sample ID}
\usage{
id_to_well(id)
}
\arguments{
\item{id}{A character string or vector of character strings formatted as an Incucyte sample ID.}
}
\value{
A character string or vector of character strings containing the well from each sample ID.
}
\description{
Utility function that returns the well number for each id.
}
\examples{
df$well <- id_to_well(df$id)
}
\keyword{id}
\keyword{tidycyte}
\keyword{well}
