% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_to_fractions.R
\name{metrics_to_fractions}
\alias{metrics_to_fractions}
\title{Calculate fractions from a set of tidycyte metrics}
\usage{
metrics_to_fractions(.df, ..., .prefix = "", .suffix = "")
}
\arguments{
\item{.df}{A data frame containing tidycyte data.}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> The masked names of the metrics to use for calculating each fraction.}

\item{.prefix}{An optional character string containing a prefix if you wish to assign the fraction to new metric names.}

\item{.suffix}{An optional character string containing a suffix if you wish to assign the fraction to new metric names.}
}
\value{
A data frame containing updated or new metrics containing fractions.
}
\description{
Accepts as arguments multiple sets of metrics and then calculates the proportion that each metric
represents out of the total sum. Effectively assumes that each metric is an object count, and
returns the proportion of objects from each set.
}
\details{
Note that in practice, this assumes that each
count is mutually exclusive. For example, if you count green, red, and yellow (green and red)
objects, then ensure that the yellow objects do not contain green-only or red-only objects.
Depending on the type of data, you may (or may not) be able to subtract these off as needed.
}
\examples{
df \%>\% metrics_to_fractions(Confluence,mAzalea,`h2-3`) # updates metrics in place
df \%>\% metrics_to_fractions(Confluence,mAzalea,`h2-3`, .prefix="fraction_", .suffix="_of_all")
df \%>\% metrics_to_fractions(mAzalea,`h2-3`, .prefix="fraction_", .suffix="_of_fluorescent")
}
\keyword{fraction}
\keyword{tidycyte}
