% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_by_metric.R
\name{normalize_by_metric}
\alias{normalize_by_metric}
\title{Normalize tidycyte metrics by other specified metrics}
\usage{
normalize_by_metric(df, .metadata)
}
\arguments{
\item{.metadata}{A data frame containing tidycyte metadata}

\item{.df}{A data frame containing tidycyte data.}
}
\value{
A data frame with normalized metrics
}
\description{
Normalizes all metrics based on the arrangement of the tidycyte metadata table. The metadata
table is expected to have a column named "\code{metric}" referring to each metric name, and a
column named "\code{metric_norm}" that contains on each row the name of another metric for normalization
or, if no normalization is desired for that metric, some other value that does not correspond
to a metric name (e.g. \code{NA}, \code{NULL}, or empty).
}
\examples{
df \%>\% normalize_by_metric(metadata)
}
\keyword{normalize}
\keyword{tidycyte}
