% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_by_time.R
\name{normalize_by_time}
\alias{normalize_by_time}
\title{Normalize tidycyte metrics by specified time points}
\usage{
normalize_by_time(df, .metadata)
}
\arguments{
\item{.metadata}{A data frame containing tidycyte metadata}

\item{.df}{A data frame containing tidycyte data.}
}
\value{
A data frame with normalized metrics
}
\description{
Normalizes all metrics based on the arrangement of the tidycyte metadata table. The metadata
table is expected to have a column named "\code{metric}" referring to each metric name, and a
column named "\code{time_norm}" that contains on each row the elapsed time to be used for normalization
or, if no normalization is desired for that metric, some other value that does not correspond
to an elapsed time value (e.g. \code{NA}, \code{NULL}, or empty).
}
\examples{
df \%>\% normalize_by_time(metadata)
}
\keyword{normalize}
\keyword{tidycyte}
