% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_incucyte_matrices.R
\name{parse_incucyte_matrices}
\alias{parse_incucyte_matrices}
\title{Read and parse Incucyte matrices}
\usage{
parse_incucyte_matrices(metadata)
}
\arguments{
\item{metadata}{A data frame containing tidycyte metadata. At a minimum should contain column names "\code{metric}" and "\code{filename}".}
}
\value{
A concatenated tidycyte dataframe with columns \code{elapsed}, \code{date_time}, \code{id}, \code{value}, \code{well}, \code{image}, \code{treatment}, \code{cell}, and \code{metric}.
}
\description{
Opens, reads, and parses all Incucyte matrix files specified in the metadata table.
}
\examples{
df <- parse_incucyte_matrices(metadata)
}
\keyword{files}
\keyword{open}
\keyword{tidycyte}
