% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_metric.R
\name{parse_metric}
\alias{parse_metric}
\title{Parse an individual incucyte metric table and return tidycyte data frame}
\usage{
parse_metric(.x)
}
\arguments{
\item{.x}{A character list corresponding to a single row of a tidycyte metadata table. At a minimum should contain names "metric" and "filename".}
}
\value{
A tidycyte dataframe with columns "\code{elapsed}", "\code{date_time}", "\code{id}", "\code{value}", "\code{well}", "\code{image}", "\code{treatment}", "\code{cell}", "\code{passage}", "\code{seeding}", and "\code{metric}".
}
\description{
Opens, reads, and parses an individual Incucyte matrix files specified as a row in the metadata table.
This function is used by \code{\link[=parse_incucyte_matrices]{parse_incucyte_matrices()}} and is not anticipated to be used by users.
}
\examples{
df <- apply(metadata,1,parse_metric) \%>\% bind_rows()
}
\keyword{file}
\keyword{open}
\keyword{tidycyte}
