% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_minimum_value_in_metric.R
\name{replace_minimum_value_in_metric}
\alias{replace_minimum_value_in_metric}
\title{Replace tidycyte metric values below threshold}
\usage{
replace_minimum_value_in_metric(.df, .min, .replace, ...)
}
\arguments{
\item{.df}{A data frame containing tidycyte data.}

\item{.min}{A numeric value specifying the minimum Values are replaced if they are below this value, but not if they are equal to this value.}

\item{.replace}{A numeric replacement value}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> The masked names of the metrics to which replacement is applied.}
}
\value{
A dataframe with adjusted metrics
}
\description{
Replaces all values of a tidycyte metric below a specified threshold with a replacement value.
}
\examples{
df \%>\% replace_minimum_value_in_metric(3,0,Confluence,mAzalea,`h2-3`)
}
\keyword{min}
\keyword{tidycyte}
