% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na_value_in_metric.R
\name{replace_na_value_in_metric}
\alias{replace_na_value_in_metric}
\title{Replace tidycyte metric values with NA values}
\usage{
replace_na_value_in_metric(.df, .replace, ...)
}
\arguments{
\item{.df}{A data frame containing tidycyte data.}

\item{.replace}{A numeric replacement value}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> The masked names of the metrics to which replacement is applied.}
}
\value{
A dataframe with adjusted metrics
}
\description{
Replaces all values of a tidycyte metric containing NA values with a replacement value.
}
\examples{
df \%>\% replace_na_value_in_metric(0,`h2-3`,mAzalea)
}
\keyword{na}
\keyword{tidycyte}
