% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_metric.R
\name{scale_metric}
\alias{scale_metric}
\title{Scale tidycyte metric}
\usage{
scale_metric(.df, .name, .to = c(0, 1), ...)
}
\arguments{
\item{.df}{A data frame containing tidycyte data.}

\item{.name}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> The masked name of the metric to normalize.}

\item{.to}{Numeric vector of the min and max final rescaled range. Defaults to \code{c(0,1)}.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Additional arguments to pass to rescale().}
}
\value{
A dataframe with rescaled metrics
}
\description{
Convenience function that uses \code{\link[=derive_metric]{derive_metric()}} and the \code{\link[scales:rescale]{scales::rescale()}} function.
By default, rescales each individual measured value from 0 to 1 using
the range from minmum to maximum value.
}
\details{
Note that this does not scale according to the mean of
the observations but to the individual measured values themselves. Hence if using \code{\link[=summary_stats]{summary_stats()}},
the maximum mean value would be expected to be less than 1. If normalization is needed after
using \code{\link[=summary_stats]{summary_stats()}}, it can be performed manually.

This function overwrites the original metric. If a new metric name is desired, the
original metric should first be copied to a new metric name. See example below.
}
\examples{
# copy to new name rather than the original name
df \%>\% derive_metric(mAzalea_scaled, mAzalea) \%>\% 
  scale_metric(mAzalea_scaled)

# rescales mAzalea metric in place, overwriting original values
df \%>\% scale_metric(mAzalea)
}
\keyword{rescale}
\keyword{tidycyte}
