% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_across_images.R
\name{sum_across_images}
\alias{sum_across_images}
\title{Sums tidycyte metrics across images}
\usage{
sum_across_images(.df, .na.rm = TRUE)
}
\arguments{
\item{.df}{A data frame containing tidycyte data.}

\item{.na.rm}{Boolean value indicating whether \code{NA} values should be removed. Defauls to TRUE.}
}
\value{
A data frame containing the metrics summed across images. The "\code{image}" field is modified to reflect the total number of images used for summing, and the "\code{id}" field is modified to note that it is an image sum.
}
\description{
Sums all metrics in a tidycyte dataframe across images. Groups images together that have the
same values of each parameter except for id, image, and value, each of which can be variable. This
effectively groups within each combination of elapsed, date_time, well, cell, treatment, and
metric.
}
\examples{
df \%>\% sum_across_images()
}
\keyword{image}
\keyword{sum}
\keyword{tidycyte}
