% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats}
\alias{summary_stats}
\title{Calculate summary statistics for tidycyte data}
\usage{
summary_stats(.data, measurevar, ..., .ci = 0.95, .na.rm = FALSE)
}
\arguments{
\item{.data}{A data frame containing tidycyte data.}

\item{measurevar}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> The masked name of the column containing tidycyte values. Typically this will be named \code{value}.}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> The masked names of the grouping variables for summary statistics. Values sharing each of these parameters will be grouped together for calculating statistics. Typically will be \code{elapsed,treatment,cell,metric}.}

\item{.ci}{Numeric value specifying the confidence interval range as a fraction. Defaults to 0.95.}

\item{.na.rm}{A boolean that indicates whether to ignore \code{NA} values. Defaults to FALSE.}
}
\description{
Gives n, mean, standard deviation, standard error of the mean, and confidence interval.
}
\examples{
df \%>\% summary_stats(value,elapsed,treatment,cell,metric,.ci = 0.95)
}
\keyword{error}
\keyword{standard}
\keyword{statistics}
\keyword{summary}
\keyword{tidycyte}
