% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoint_norm_group.R
\name{timepoint_norm_group}
\alias{timepoint_norm_group}
\title{Normalize an individual tidycyte metric data frame by time}
\usage{
timepoint_norm_group(df, t0)
}
\arguments{
\item{df}{A data frame of an individual tidycyte dataset. At a minimum should contain column names "value" and "elapsed".}

\item{t0}{A numeric value specifying an individual elapsed time.}
}
\value{
A dataframe with the value column normalized by the value where elapsed equals t0.
}
\description{
Normalizes an individual tidycyte data frame based on a specified timem value.
This function is used by \code{\link[=normalize_by_time]{normalize_by_time()}} to normalize pre-grouped data, and is not
anticipated to be used by users.
}
\examples{
df \%>\%
  group_by(across(-c(date_time,elapsed,id,value))) \%>\%
  nest() \%>\%
  mutate(data = map(data, timepoint_norm_group, t0[[metric]])) \%>\%
  unnest(cols = c(data)) \%>\%
  ungroup()
}
\keyword{normalize}
\keyword{tidycyte}
\keyword{timepoint}
